refactor topo logics
