+++ include/mori/application/topology/topology.hpp
@@ -1,3 +1,7 @@
 #pragma once
 
+#include "mori/application/topology/gpu.hpp"
+#include "mori/application/topology/net.hpp"
+#include "mori/application/topology/node.hpp"
+#include "mori/application/topology/pci.hpp"
 #include "mori/application/topology/system.hpp"
